requirements.txt
pywebview==6.1
pythonnet==3.0.5
clr_loader==0.2.10
pyinstaller==6.17.0
pillow==12.1.0
